// src/types/omnivoice-api.d.ts
/**
  * This file is auto-generated by scripts/sync-api.ts
  * Do not edit manually. Run `npm run sync-api` to regenerate.
  *
  * Generated from: https://nvovk-2004--omnivoice-tts-omnivoicetts-serve.modal.run/openapi.json
  * Generated at: 2026-09-25T16:23:10.410Z
  */
export interface paths {
  '/generate': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /** Generate Speech */
    post: operations['generate_speech_generate_post']
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    /** HTTPValidationError */
    HTTPValidationError: {
      /** Detail */
      detail?: components['schemas']['ValidationError'][]
    }
    /**
         * TTSRequest
         * @description Request model for text-to-speech generation.
         */
    TTSRequest: {
      /** Prompt */
      prompt: string
      /** Voice Key */
      voice_key: string
      /** Ref Text */
      ref_text: string
      /**
             * Language
             * @default uk
             */
      language: string | null
      /**
             * Speed
             * @default 1
             */
      speed: number
      /**
             * Num Step
             * @default 32
             */
      num_step: number
      /**
             * Guidance Scale
             * @default 2
             */
      guidance_scale: number
    }
    /** ValidationError */
    ValidationError: {
      /** Location */
      loc: (string | number)[]
      /** Message */
      msg: string
      /** Error Type */
      type: string
    }
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}
export type $defs = Record<string, never>;
export interface operations {
  generate_speech_generate_post: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        'application/json': components['schemas']['TTSRequest']
      }
    }
    responses: {
      /** @description Successful Response */
      200: {
        headers: Record<string, unknown>
        content: {
          'application/json': unknown
          'audio/wav': unknown
        }
      }
      /** @description Validation Error */
      422: {
        headers: Record<string, unknown>
        content: {
          'application/json': components['schemas']['HTTPValidationError']
        }
      }
    }
  }
}